# `getSkipSegments` rework