--- a/src/routes/getSkipSegments.js
+++ b/src/routes/getSkipSegments.js
@@ -113,17 +113,17 @@ function getWeightedRandomChoice(choices, weights, amountOfChoices) {
 function getVoteOrganisedSponsorTimes(sponsorTimes, votes, UUIDs) {
   //create groups of sponsor times that are similar to eachother
   const groups = []
-  sponsorTimes.forEach(([sponsorStart, sponsorEnd], i) => {
+  sponsorTimes.forEach(([startTime, endTime], i) => {
       //find a group that overlaps with the current segment
       //sponsorTimes are sorted by their startTime so there should never be more than 1 similar group
-      const similarGroup = groups.find(group => group.start < sponsorEnd && sponsorStart < group.end)
+      const similarGroup = groups.find(group => group.start < endTime && startTime < group.end)
       //add the sponsor to that group or create a new group if there aren't any
       if (similarGroup === undefined) {
-          groups.push({ start: sponsorStart, end: sponsorEnd, sponsors: [i] })
+          groups.push({ start: startTime, end: endTime, sponsors: [i] })
       } else {
           similarGroup.sponsors.push(i)
-          similarGroup.start = Math.min(similarGroup.start, sponsorStart)
-          similarGroup.end = Math.max(similarGroup.end, sponsorEnd)
+          similarGroup.start = Math.min(similarGroup.start, startTime)
+          similarGroup.end = Math.max(similarGroup.end, endTime)
       }
   })