This is a rewrite of getSkipSegments.js to simplify the code and hopefully improve performance. The behaviour remains the same as before, only the implementation changes.

The main performance improvement comes from the fact that `handleGetSegments` will now only hash the user IP when it's necessary (when it needs to filter out the shadow hidden segments) instead of doing it everytime, no matter what.

Here are a few measurements for a `handleGetSegments` call for a video that doesn't have shadow hidden segments (ie no need to hash the IP, ie the best case scenario):

Before:
```
20.212ms
15.036ms
12.652ms
11.665ms
10.165ms
10.242ms
10.088ms
9.747ms
13.810ms
13.782ms
```

After:
```
1.467ms
0.315ms
0.296ms
0.292ms
0.275ms
0.270ms
0.294ms
0.270ms
0.283ms
0.278ms
```