@ajayyy 2020-06-07
I didn't realize hashing took that much time